refactor(components/containers): update action/interface imports
