+++ client/src/components/Dashboard/UserLessonList/UserLessonList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { IUserLesson, IUserLessonListProps } from '../../../interfaces';
 import { AppState } from '../../../store/configureStore';
+import { IUserLesson, IUserLessonListProps } from '../../../interfaces';
 import { UserLesson } from '../../../components';
 
 export default function UserLessonList(props: IUserLessonListProps): JSX.Element {
